translib: split transceiver getters into pfm_app_transceiver.go
Address review comment on PR #201: pfm_app.go had grown to roughly
2100 lines, dominated by the OpenConfig Platform Transceiver
getters that were appended for #201. Per the maintainer's note that
breaking the new logic into separate files is acceptable, move the
transceiver Db-struct types and getCompTransceiver* getters into a
new translib/pfm_app_transceiver.go in the same package.

What stays in pfm_app.go:
  - PlatformApp struct, init / translate* / process*
  - The four map-lookup helpers (getTransceiverInfoEntry,
    getTransceiverDomSensorEntry, getTransceiverDomThresholdEntry,
    getApplPortEntry) so callers across both files can share them
  - getCompStateDbObj / getCompStateFromDb (System Component state)
  - doGetPlatformInfo, which continues to be the single dispatch
    entry point and now calls into the transceiver getters
    located in the new file

What moves to pfm_app_transceiver.go:
  - CompTransceiverStateDb / CompTransceiverStateSupplyVoltageDb /
    CompTransceiverPhysicalChannelState{Laser{Temperature,
    BiasCurrent}, OutputPower, InputPower}Db /
    CompTransceiverThresholdStateDb
  - getCompTransceiverStateDbObj / getCompTransceiverStateFromDb
  - getCompTransceiverStateSupplyVoltage{DbObj, FromDb}
  - getCompTransceiverPhysicalChannelStateLaserTemperature{DbObj, FromDb}
  - getCompTransceiverPhysicalChannelStateOutputPower{DbObj, FromDb}
  - getCompTransceiverPhysicalChannelStateInputPower{DbObj, FromDb}
  - getCompTransceiverPhysicalChannelStateLaserBiasCurrent{DbObj, FromDb}
  - getCompTransceiverThresholdState{DbObj, FromDb}

This is purely a file-organisation change: no symbol is renamed,
no logic is altered. pfm_app.go drops from ~1990 to ~1440 lines,
and the new file is ~580 lines.

Drop the now-unused math, regexp and (already-unused) reflect-adjacent
imports from pfm_app.go's import block. pfm_app_transceiver.go has
its own minimal import block (fmt, math, regexp, strconv, strings,
ocbinds, glog).

doGetPlatformInfo itself remains long; further extraction by URI
pattern is out of scope for this patch and will be a follow-up.
